feat(mcp): add safe snapshot parity and warn on untrusted content
