+++ docs/mcp.md
@@ -5,6 +5,11 @@ PinchTab includes a native [Model Context Protocol (MCP)](https://modelcontextpr
 > [!WARNING]
 > The MCP server is part of PinchTab's privileged control plane. It is intended for trusted operators and trusted agent systems only. Do not expose it to untrusted users, untrusted client systems, or the public internet. If you are unsure how to secure a non-local deployment, review [Security](guides/security.md) and use the private security contact path in `SECURITY.md` before exposing the service.
 
+> [!CAUTION]
+> By default, PinchTab's IDPI posture is meant to keep MCP browsing local-only until you deliberately widen it. Expanding MCP use to non-local or non-trusted domains is a security-reducing choice.
+>
+> When MCP tools read page content from wider domains, treat `pinchtab_snapshot` and `pinchtab_get_text` output as untrusted data, not instructions. Hostile pages can contain prompt-injection content, poisoned text, or other material that should never be treated as operator guidance. Review [Security](guides/security.md#idpi) before relaxing domain restrictions.
+
 ## Quick Start
 
 1. Start PinchTab in server or bridge mode:
@@ -161,9 +166,15 @@ Common selector forms:
 
 The normal MCP browser loop is:
 
-1. `pinchtab_navigate`
-2. `pinchtab_snapshot`
-3. `pinchtab_click` / `pinchtab_type` / other action tools
-4. `pinchtab_wait_*` or `pinchtab_network` when needed
+1. Call `pinchtab_navigate` with a `url`
+2. Call `pinchtab_snapshot` to inspect page structure and collect refs
+3. Call `pinchtab_click`, `pinchtab_type`, or other action tools with structured arguments
+4. Call `pinchtab_wait_*` or `pinchtab_network` when needed
+
+`pinchtab_snapshot` supports MCP-safe output controls:
+
+- `compact=true` or `format="compact"` for the most token-efficient text snapshot
+- `format="text"` for the full text snapshot
+- `noAnimations=true` to reduce animation noise before capture
 
 For full parameter details, see [MCP Tool Reference](./reference/mcp-tools.md).
