+++ internal/mcp/tools.go
@@ -16,10 +16,12 @@ func allTools() []mcp.Tool {
 			mcp.WithString("tabId", mcp.Description("Target tab ID")),
 			mcp.WithBoolean("interactive", mcp.Description("Only interactive elements (buttons, links, inputs)")),
 			mcp.WithBoolean("compact", mcp.Description("Compact format (most token-efficient)")),
+			mcp.WithString("format", mcp.Description("Output format: 'compact' or 'text'")),
 			mcp.WithBoolean("diff", mcp.Description("Only changes since last snapshot")),
 			mcp.WithString("selector", mcp.Description("Unified selector to scope the snapshot (CSS, XPath, text, or ref)")),
 			mcp.WithNumber("maxTokens", mcp.Description("Maximum estimated tokens in response (e.g. 300)")),
 			mcp.WithNumber("depth", mcp.Description("Maximum tree depth (e.g. 3)")),
+			mcp.WithBoolean("noAnimations", mcp.Description("Disable animations before capturing the snapshot")),
 		),
 		mcp.NewTool("pinchtab_screenshot",
 			mcp.WithDescription("Take a screenshot of the current page"),
@@ -31,6 +33,7 @@ func allTools() []mcp.Tool {
 			mcp.WithDescription("Extract readable text content from the current page"),
 			mcp.WithString("tabId", mcp.Description("Target tab ID")),
 			mcp.WithBoolean("raw", mcp.Description("Return raw text without formatting")),
+			mcp.WithString("format", mcp.Description("Response format: 'text'/'plain' for plain text, default JSON envelope")),
 			mcp.WithNumber("maxChars", mcp.Description("Maximum characters in response (e.g. 3000)")),
 		),
 
@@ -39,6 +42,7 @@ func allTools() []mcp.Tool {
 			mcp.WithDescription("Click an element by selector (ref from snapshot e.g. 'e5', CSS e.g. 'css:#btn', XPath e.g. 'xpath://button', text e.g. 'text:Submit', or semantic e.g. 'find:login button')"),
 			mcp.WithString("selector", mcp.Required(), mcp.Description("Unified selector: ref (e.g. 'e5'), CSS ('css:#btn' or '#btn'), XPath ('xpath://button'), text ('text:Submit'), or semantic ('find:login button')")),
 			mcp.WithString("ref", mcp.Description("(deprecated) Element ref from snapshot — use 'selector' instead")),
+			mcp.WithBoolean("waitNav", mcp.Description("Wait for navigation after the click when the element triggers a page change")),
 			mcp.WithString("tabId", mcp.Description("Target tab ID")),
 		),
 		mcp.NewTool("pinchtab_type",
